Common.System: removed unused 'getInstallDir'
The function was already commented out in its only use case in
'Compiler.Options', so the compiler's functionality is not all
affected.

The function was extremely platform dependent and did not work
in a meaninful manner at least in my Linux environment.

Even worse, it introduced a dependency on Windows-only C code.